Integrating changes into one map
First pass of pulling in all changes into one map.
Deleted placeholder prefabs.
Created new prefabs for the wisps and the towers.
Added a "placement helper" - a plane that has the map texture on it. In
game, the plane will disappear.
Added a new component that will "ground" any entity it's attached to on
startup. (This should help place things in the map, since their
positions will be fixed up when the game starts)